Add map markers for subregions
When the user activates a region, markers are added for any subregions.
When that region is unactivated, the markers are removed.

The style comes from the prototypes. Also, the leaflet CSS was moved to
the header section above the local stylesheet so that the local
stylesheet has precedent.  In order to replicate the style in the mock-ups,
where the marker icon is used as the popup tip, a second marker must be
added to a custom pane that is above the popup pane. This is because markers
are rendered under the popups by default.

Refs #13