Refactor use of models and views
Refactors the existing code to make better use of Backbone models and
views, and to conform to Backbone best practices. These changes reduce
the need for custom state management and rendering code, and take
advantage of some built-in Backbone features.

This is all in anticipation of implementing fly-to transitions and
detail views for subregions, which will require cordination between the
map and sidebar. This is made easier by instantiating the regions and
subregions as models, which have state that can be accessed throughout
the app.