Add fly-to transitions for subregions
When a user clicks on a subregion in the sidebar, the description of the
subregion is animated in below the subregion button. Also, the map does
a fly-to transition to the subregion location and opens the subregion popup.
Selecting another subregion in the same region closes the initial popup
and transitions to the selected region. Exiting the region closes the
popup, removes the markers, and transitions back to the initial map state.

Refs #32